requirements.txt
# av  # required but typically not installable via pip, try `conda install -c conda-forge av`
ephyviewer>=1.3.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
neo>=0.7.2
numpy
packaging
pandas
pyqt5
pyyaml
quantities
requests
scipy
setuptools
tqdm
